Fix Esp32/Rp2040 `analogRead()` function (#2973)
This PR fixes a couple of bugs revealed by #2971 where `analogRead` fails when running on the `esp32c3`.

Instead of a fixed definition for `A0` in `pins_arduino.h`, we use the definitions provided by the ESP-IDF which vary depending on the specific SoC being targetted.

In addition, the call to `analogRead` failed because the ADC unit number was off by one (zero-based, not 1-based). I guess the esp32 doesn't care, but the esp32c3 does. The `analogRead` function can be simplified to use the IDF for this as we no longer require support for IDF prior to version 5.

See https://github.com/SmingHub/Sming/issues/2971#issuecomment-3290721812